fix: move routing.consts to routing folder
