+++ src/routing/Routers.test.tsx
@@ -4,9 +4,9 @@ import { SM_DATASOURCE } from 'test/fixtures/datasources';
 import { type CustomRenderOptions, render } from 'test/render';
 
 import { ROUTES } from 'types';
+import { PLUGIN_URL_PATH } from 'routing/constants';
 import { InitialisedRouter, UninitialisedRouter } from 'routing/Routers';
 import { getRoute } from 'routing/utils';
-import { PLUGIN_URL_PATH } from 'components/Routing.consts';
 
 function renderInitialisedRouting(options?: CustomRenderOptions) {
   return render(<InitialisedRouter />, options);
